test(python): split selective json parser tests by contract
